Trait halo2_backend::poly::commitment::ParamsProver

pub trait ParamsProver<C: CurveAffine>: Params<C> {
    // Required methods
    fn new(k: u32) -> Self;
    fn commit(
        &self,
        engine: &impl MsmAccel<C>,
        poly: &Polynomial<C::ScalarExt, Coeff>,
        r: Blind<C::ScalarExt>
    ) -> C::CurveExt;
}

Parameters for circuit synthesis and prover parameters.

Required Methods

fn new(k: u32) -> Self

Returns new instance of parameters

fn commit( &self, engine: &impl MsmAccel<C>, poly: &Polynomial<C::ScalarExt, Coeff>, r: Blind<C::ScalarExt> ) -> C::CurveExt

This computes a commitment to a polynomial described by the provided slice of coefficients. The commitment may be blinded by the blinding factor r.

Object Safety

This trait is not object safe.