Module halo2_backend::multicore

Structs

Scope

Represents a fork-join scope which can be used to spawn any number of tasks. See scope() for more information.

Traits

IndexedParallelIterator

An iterator that supports “random access” to its data, meaning that you can split it at arbitrary indices and draw data from those points.

IntoParallelIterator

IntoParallelIterator implements the conversion to a ParallelIterator.

IntoParallelRefIterator

IntoParallelRefIterator implements the conversion to a ParallelIterator, providing shared references to the data.

IntoParallelRefMutIterator

IntoParallelRefMutIterator implements the conversion to a ParallelIterator, providing mutable references to the data.

ParallelIterator

Parallel version of the standard iterator trait.

ParallelSliceMut

Parallel extensions for mutable slices.

TryFoldAndReduce

Functions

current_num_threads

Returns the number of threads in the current registry. If this code is executing within a Rayon thread-pool, then this will be the number of threads for the thread-pool of the current thread. Otherwise, it will be the number of threads for the global thread-pool.

join

Takes two closures and potentially runs them in parallel. It returns a pair of the results from those closures.

scope

Creates a “fork-join” scope s and invokes the closure with a reference to s. This closure can then spawn asynchronous tasks into s. Those tasks may run asynchronously with respect to the closure; they may themselves spawn additional tasks into s. When the closure returns, it will block until all tasks that have been spawned into s complete.